--- lib/csharp/microsoft.bot.builder.skills/Microsoft.Bot.Builder.Skills/Contextual/IUserContext.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Microsoft.Bot.Builder.Skills.Contextual
-{
-    public interface IUserContextResolver
-    {
-
-
-    }
-}
