+++ lib/csharp/microsoft.bot.builder.skills/Microsoft.Bot.Builder.Skills/Contextual/UserContextResolver.cs
@@ -2,22 +2,38 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Threading.Tasks;
 
 namespace Microsoft.Bot.Builder.Skills.Contextual
 {
     public class UserContextResolver
     {
-        private IUserContextResolver _contextResolver;
+        private IContextResolver _contextResolver;
+        private UserStateContextResolver _userStateContextResolver;
 
-        public UserContextResolver(UserInfoState userInfo, IUserContextResolver contextResolver = null)
+        public UserContextResolver(UserInfoState userInfo, IContextResolver contextResolver = null)
         {
-            this._contextResolver = contextResolver;
+            _contextResolver = contextResolver;
+            _userStateContextResolver = new UserStateContextResolver(userInfo);
         }
 
-        //public string GetResolvedContext(string context)
-        //{
+        public async Task<IList<string>> GetResolvedContactAsync(RelatedEntityInfo relatedEntityInfo)
+        {
+            // Take result as following priority:
+            // 1. Injection context resolver
+            if (_contextResolver != null)
+            {
+                var resolvedContact = await _contextResolver.GetResolvedContactAsync(relatedEntityInfo);
 
-        //}
+                if (resolvedContact != null)
+                {
+                    return resolvedContact;
+                }
+            }
 
+            // 2. User state context resolver
+            var resolvedUserStateContact = await _userStateContextResolver.GetResolvedContactAsync(relatedEntityInfo);
+            return resolvedUserStateContact;
+        }
     }
 }
